Arduino: Finally Operating Correctly
I don't know how I didn't catch this bug earlier.
The boolean timeout value in the polling while loops needed to be
inverted (!).
Aparently the network packets pass faster then the ESP processes,
because this caused only intermittent bugs.
In debugging, I changed one of the reads to a different method. Not sure
which I will stick with.